backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
apscheduler==3.10.4
pandas==2.1.4
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
twilio==8.11.0
httpx==0.26.0
python-dateutil==2.8.2
